pyproject.toml
[project]
name = "ai-powered-financial-intelligence-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.136.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
